--- image-compressor/README.md
@@ -1,4 +1,0 @@
-# ImageCompressor
-An extension for MIT App Inventor 2.
-Created by: Aril Ogai
-Compiled by: FAST
