+++ src/AndroidManifest.xml
