+++ src/com/bosonshiggs/imagecompressor/ImageCompressor.java
@@ -1,30 +1,33 @@
 package com.bosonshiggs.imagecompressor;
 
-import com.bosonshiggs.imagecompressor.helpers.PathUtils;
+import java.io.File;
+import java.io.FileOutputStream;
 
-import com.google.appinventor.components.annotations.*;
-import com.google.appinventor.components.runtime.*;
-import com.google.appinventor.components.common.ComponentCategory;
-
-import android.net.Uri;
-import android.content.Context;
+import android.app.Activity;
 import android.content.Intent;
 import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
+import android.media.MediaScannerConnection;
+import android.net.Uri;
 import android.os.Environment;
 import android.provider.MediaStore;
-import android.media.MediaScannerConnection;
 import android.util.Log;
-import android.app.Activity;
 
-import java.io.File;
-import java.io.FileOutputStream;
+import com.google.appinventor.components.annotations.Asset;
+import com.google.appinventor.components.annotations.DesignerComponent;
+import com.google.appinventor.components.annotations.SimpleEvent;
+import com.google.appinventor.components.annotations.SimpleFunction;
+import com.google.appinventor.components.annotations.SimpleProperty;
+import com.google.appinventor.components.runtime.ActivityResultListener;
+import com.google.appinventor.components.runtime.AndroidNonvisibleComponent;
+import com.google.appinventor.components.runtime.ComponentContainer;
+import com.google.appinventor.components.runtime.EventDispatcher;
 
 @DesignerComponent(
-    version = 7,
-    versionName = "1.0",
-    description = "An extension for compressing images to JPEG, PNG, or WebP.",
+    version = 7, 
+    versionName = "1.0", 
+    description = "An extension for compressing images to JPEG, PNG, or WebP.", 
     iconName = "icon.png"
 )
 public class ImageCompressor extends AndroidNonvisibleComponent implements ActivityResultListener {
@@ -37,19 +40,18 @@ public class ImageCompressor extends AndroidNonvisibleComponent implements Activ
     public ImageCompressor(ComponentContainer container) {
         super(container.$form());
         this.container = container;
-        form.registerForActivityResult(this);
+        container.$form().registerForActivityResult(this);
     }
 
-    @SimpleProperty(
-        description = "Enable or disable logging to Logcat"
-    )
+    @SimpleProperty(description = "Enable or disable logging to Logcat")
     public void LogEnabled(boolean enabled) {
         this.flagLog = enabled;
     }
 
     @SimpleFunction(description = "Converts a content URI to a file system path.")
     public String PathFromUri(String stringUri) {
-        if (flagLog) Log.d("ImageCompressor", "Converting URI to path: " + stringUri);
+        if (flagLog)
+            Log.d("ImageCompressor", "Converting URI to path: " + stringUri);
         if (stringUri == null || stringUri.isEmpty()) {
             return "";
         }
@@ -61,15 +63,16 @@ public class ImageCompressor extends AndroidNonvisibleComponent implements Activ
                 return cursor.getString(columnIndex);
             }
         } catch (Exception e) {
-            if (flagLog) Log.e("ImageCompressor", "Error converting URI to path", e);
+            if (flagLog)
+                Log.e("ImageCompressor", "Error converting URI to path", e);
         }
         return stringUri;
     }
 
     @SimpleFunction(description = "Compresses the image at the given path to the specified format and quality.")
-    @SuppressWarnings("deprecation")
     public void CompressImage(@Asset final String imagePath, final String imageName, final int quality, final int id) {
-        if (flagLog) Log.d("ImageCompressor", "Starting image compression: " + imagePath);
+        if (flagLog)
+            Log.d("ImageCompressor", "Starting image compression: " + imagePath);
         new Thread(new Runnable() {
             @Override
             public void run() {
@@ -106,23 +109,26 @@ public class ImageCompressor extends AndroidNonvisibleComponent implements Activ
                         // Always use WEBP_LOSSY to avoid deprecated constant
                         bitmap.compress(Bitmap.CompressFormat.WEBP_LOSSY, quality, fos);
                     } else {
-                        if (flagLog) Log.e("ImageCompressor", "Unsupported file format: " + imageType);
+                        if (flagLog)
+                            Log.e("ImageCompressor", "Unsupported file format: " + imageType);
                         fos.close();
                         return;
                     }
 
                     fos.close();
-                    MediaScannerConnection.scanFile(form, new String[]{ imageFile.getAbsolutePath() }, null, null);
+                    MediaScannerConnection.scanFile(form, new String[] { imageFile.getAbsolutePath() }, null, null);
 
                     container.$form().runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
                             ImageSaved(id, imageFile.getAbsolutePath());
-                            if (flagLog) Log.i("ImageCompressor", "Image compressed and saved: " + imageFile.getAbsolutePath());
+                            if (flagLog)
+                                Log.i("ImageCompressor", "Image compressed and saved: " + imageFile.getAbsolutePath());
                         }
                     });
                 } catch (final Exception e) {
-                    if (flagLog) Log.e("ImageCompressor", "Error compressing and saving image", e);
+                    if (flagLog)
+                        Log.e("ImageCompressor", "Error compressing and saving image", e);
                     container.$form().runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
